页面的博客的数据已做好，可以对接，传出博客的分类，和博客前五条的数据，和pages

+++ blogModule/src/com/bdqn/blog/controller/BlogController.java
@@ -4,6 +4,7 @@ import com.bdqn.blog.pojo.Blog;
 import com.bdqn.blog.pojo.BlogGenre;
 import com.bdqn.blog.server.BlogGenreServer;
 import com.bdqn.blog.server.BlogService;
+import com.bdqn.blog.utils.PageSupport;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -29,6 +30,7 @@ public class BlogController {
     private BlogService blogService;
     @Resource
     private BlogGenreServer BlogGenreServer;
+    private PageSupport pageSupport=new PageSupport();
     //增加博客
     @RequestMapping(value = "/add" ,method = RequestMethod.POST)
     public String addBolg(@RequestParam String title,@RequestParam String contentPath){
@@ -62,18 +64,37 @@ public class BlogController {
     @RequestMapping("/removeBlog")
     public  String removeBlog(Integer id){
         if(id !=null){
-            blogService.removeBlog(id);
+            Blog blog=blogService.selectByBid(id);
+            if(blog!=null){
+                blogService.removeBlog(id);
+            }
+
             return "blogWelcome";
         }
         return null;
     }
-    //查询博客
+    /**
+     *  查询博客
+     *
+     *  传入  uid用户ID   ， like  title 标题   ，pageNo
+     */
+
     @RequestMapping("/selectBlog")
-    public String selectBlog(Model Model , @RequestParam(value = "pageNo" ,required = false) int pageNo,
+    public String selectBlog(Model Model , @RequestParam(value = "pageNo" ,required = false) Integer pageNo,
                              @RequestParam(value = "uid" ,required = false)   int uid,
                              @RequestParam(value = "title" ,required = false)    String title){
         List<BlogGenre> BlogGenres=BlogGenreServer.getBlogGenreAll();
-        List<Blog> blogs= blogService.selectAllBlog(uid,title,pageNo,3);
+
+        if(pageNo!=null){
+            pageSupport.setCurrentPageNo(pageNo);
+        }
+        List<Blog> blogs= blogService.selectAllBlog(uid,title,pageSupport.getCurrentPageNo()-1,pageSupport.getPageSize());
+
+        pageSupport.setTotalCount(blogService.totalCount(uid,title) );
+
+        pageSupport.setTotalPageCountByRs();
+
+        Model.addAttribute("pages",pageSupport);
         Model.addAttribute("blogs",blogs);
         Model.addAttribute("BlogGenres",BlogGenres);
         return  "blogWelcome";
